requirements.txt
jaseci
byllm
gitpython
tree_sitter
python-dotenv
rich
